+++ app/components/views/TicketsPage/MyTicketsTab/TicketsCardList.js
@@ -34,7 +34,7 @@ class TicketsCardList extends React.Component {
       })}
     </div>);
   }
-};
+}
 
 export default TicketsCardList;
